+++ demo-root/demo-api/src/main/java/org/sklsft/demo/api/interfaces/reference/localization/base/RegionBaseService.java
@@ -26,7 +26,7 @@ public static final String GET_LIST_URL = "/region/list";
 /**
  * load object list from country
  */
-List<RegionBasicView> loadListFromCountry (Long countryId);
+List<RegionBasicView> loadListFromCountry (Short countryId);
 public static final String GET_REGION_LIST_fROM_COUNTRY_URL = "/country/{countryId}/region/list";
 
 /**
@@ -38,13 +38,13 @@ public static final String SCROLL_URL = "/region/scroll";
 /**
  * scroll object list from country
  */
-ScrollView<RegionBasicView> scrollFromCountry (Long countryId, ScrollForm<RegionFilter, RegionSorting> form);
+ScrollView<RegionBasicView> scrollFromCountry (Short countryId, ScrollForm<RegionFilter, RegionSorting> form);
 public static final String SCROLL_REGION_fROM_COUNTRY_URL = "/country/{countryId}/region/scroll";
 
 /**
  * load object
  */
-RegionFullView load(Long id);
+RegionFullView load(Integer id);
 public static final String GET_URL = "/region/{id}";
 
 /**
@@ -62,31 +62,31 @@ public static final String GET_NEW_URL = "/region/new";
 /**
  * save object
  */
-Long save(RegionForm regionForm);
+Integer save(RegionForm regionForm);
 public static final String SAVE_URL = "/region";
 
 /**
  * save object from parent Country
  */
-Long saveFromCountry(Long countryId, RegionForm regionForm);
+Integer saveFromCountry(Short countryId, RegionForm regionForm);
 public static final String SAVE_FROM_COUNTRY_URL = "/country/{countryId}/region";
 
 /**
  * update object
  */
-void update(Long id, RegionForm regionForm);
+void update(Integer id, RegionForm regionForm);
 public static final String UPDATE_URL = "/region/{id}";
 
 /**
  * delete object
  */
-void delete(Long id);
+void delete(Integer id);
 public static final String DELETE_URL = "/region/{id}";
 
 /**
  * delete object list
  */
-void deleteList(List<Long> idList);
+void deleteList(List<Integer> idList);
 public static final String DELETE_LIST_URL = "/region/delete";
 
 }
